Fix build issues: cmake_minimum_required policy and float-to-double promotion

+++ src/modules/simulation/gz_bridge/GZMixingInterfaceWheel.cpp
@@ -80,7 +80,7 @@ bool GZMixingInterfaceWheel::updateOutputs(bool stop_wheels, uint16_t outputs[MA
 			// Offsetting the output allows for negative values despite unsigned integer to reverse the wheels
 			static constexpr float output_offset = 100.0f;
 			float scaled_output = (float)outputs[i] - output_offset;
-			wheel_velocity_message.set_velocity(i, scaled_output);
+			wheel_velocity_message.set_velocity(i, static_cast<double>(scaled_output));
 		}
